# conf/openapi.yaml
openapi: "3.0.3"
info:
  title: Deposplit API
  version: "0.1"
  description: |
    REST API for the Deposplit secret-sharing Web app/service.

    ## Authentication

    Every request must carry three headers:

    | Header | Content |
    |---|---|
    | `X-Deposplit-Verify-Key` | Caller's public verify key (base64url — see `PublicKey` schema) |
    | `X-Deposplit-Nonce` | Per-request unique string in the form `<unix-ms>.<random>` — server rejects requests whose embedded timestamp is more than 5 minutes old |
    | `X-Deposplit-Signature` | Base64url Ed25519 signature (RFC 8032) |

    The signature covers the canonical string:

    ```
    nonce || "\n" || UPPERCASE(method) || "\n" || path_with_query || "\n" || hex(SHA-256(body))
    ```

    where `body` is the empty string for requests without a body.
    The Web app/service verifies the signature against the verify key supplied in `X-Deposplit-Verify-Key`.
    No pre-registration is required — the keypair is the identity.

    ## Payload signatures

    Two additional Ed25519 signatures ride with every `ShareRequest` row, independent of the
    transport-auth signature above. The transport signature authenticates the HTTP caller for one
    specific call and is never persisted — it gives a later reader of a row (via `GET`) nothing to
    re-verify authorship against. `senderSignature` and `recipientSignature` close that gap: any
...
    through a relay other than deposplit.com, e.g. a self-hosted instance) safe — a passive
...
servers:
...
security:
  - ed25519Auth: []
components:
  securitySchemes:
    ed25519Auth:
      type: apiKey
      in: header
      name: X-Deposplit-Signature
      description: >
        Must accompany X-Deposplit-Verify-Key and X-Deposplit-Nonce headers.
        See top-level description for the canonical signing string.
  schemas:
    SecretId:
      type: string
      format: uuid
      description: UUID generated by the sender at split time; shared across all n shares of a secret.
    PublicKey:
      type: string
      description: >
        Base64url-encoded signing-algorithm public key (32 bytes, no padding, for the current
        cipher suite `ed25519+x25519-v1` — see `CipherSuite`) — used for API authentication and as
        the routing identity. Not pinned to an exact length; a future cipher suite may use a
        different key size.
...
    ShareTransactionType:
      type: string
      enum: [deposit, retrieval, removal, inventory]
      description: >
        `deposit`   — sender deposits an encrypted share for the recipient to approve and collect.
        `retrieval` — sender wants the share bytes back from the recipient (requires approval + ciphertext).
        `removal`   — sender wants the recipient to delete their local copy (requires approval).
        `inventory` — a holder pushes a metadata-only report about a share of theirs back
        to its owner, so a recovering owner (fresh device, empty local state) can rebuild their
        records after losing their old identity's key.
        Unlike the other three, this is **not** consent-gated: it self-approves at
        creation (no `pending` phase), never carries `ciphertext`, and requires `k`/`n`/`mimeType`.
        The recipient polls for it and deletes the row once consumed.
    ShareRequestState:
      type: string
      enum: [pending, approved, denied, withdrawn]
      description: >
        `withdrawn` applies only to `deposit` rows: set via `POST /share-requests/withdraw`
        when a holder unilaterally stops holding a share, instead of the row being hard-deleted, so
        the sender's next poll can observe the tombstone. Best-effort and fire-and-forget — the
        relay may still garbage-collect the row at any time, so its absence must never be read as a
        signal; only an *observed* `withdrawn` row counts.
    ShareRequest:
      description: >
        A self-describing share request row. All four transaction types (deposit, retrieval,
        removal, inventory) use the same flat structure. `ciphertext`, `shareId`, `k`, `n`, and
        `mimeType` semantics depend on `transactionType` — see field descriptions.
      type: object
      required: [id, secretId, senderKey, recipientKey, label, secretCreatedAt, transactionType, state, requestedAt, senderSignature]
      properties:
        id:
          type: string
          format: uuid
        secretId:
          $ref: '#/components/schemas/SecretId'
        senderKey:
          $ref: '#/components/schemas/PublicKey'
        recipientKey:
          $ref: '#/components/schemas/PublicKey'
        label:
          type: string
          description: Human-readable label for the secret (display only).
        secretCreatedAt:
          type: string
          format: date-time
          description: Client-supplied timestamp of when the secret was created on the sender's device.
        transactionType:
          $ref: '#/components/schemas/ShareTransactionType'
        state:
          $ref: '#/components/schemas/ShareRequestState'
        shareId:
          type: string
          format: uuid
          nullable: true
          description: >
            For `retrieval` and `removal` rows: the `id` of the originating `deposit` request,
            supplied by the client and stored opaquely by the relay. `null` for `deposit` and
            `inventory` rows (both are roots).
        requestedAt:
          type: string
          format: date-time
        respondedAt:
          type: string
          format: date-time
          nullable: true
          description: >
            Set when the recipient responds via PATCH. For `inventory` this is set
            immediately at creation (self-approved, no PATCH phase).
        ciphertext:
          type: string
          format: byte
          nullable: true
          maxLength: 524288
          description: >
            At most 512 KiB decoded; a larger one is refused with 413. Semantics differ by type:

            `deposit`: populated by the sender at creation (opaque encrypted share bytes);
            delivered to the recipient in the PATCH response on approval and cleared from the relay.
            Not returned on GET after delivery.

            `retrieval`: absent at creation; supplied by the recipient in the PATCH body on approval;
            stored temporarily until the sender collects it via GET.

            `removal`: always null.

            `inventory`: always null — this type never carries share bytes.
        k:
          type: integer
          nullable: true
          description: >
            SSS threshold. Populated for `deposit` and `inventory` (required, `2 <= k <= n
            <= 255`); `null` for `retrieval`/`removal`.
        n:
          type: integer
          nullable: true
          description: >
            SSS share count. Populated for `deposit` and `inventory` (required, `2 <= k <=
            n <= 255`); `null` for `retrieval`/`removal`.
        mimeType:
          type: string
          nullable: true
          description: >
            The sender's claim about what the secret is, so the recipient can decide how to render
            it once reconstructed. Populated for `deposit` and `inventory`; `null` for
            `retrieval`/`removal`. Sender-supplied and never verified against the payload — the
            relay routes it opaquely, exactly as it does `label`.
        senderSignature:
          type: string
          description: >
            Base64url Ed25519 signature by senderKey — see "Payload signatures" above.
            Set at creation, never cleared.
        recipientSignature:
          type: string
          nullable: true
          description: >
            Base64url Ed25519 signature by recipientKey — see "Payload signatures" above.
            Null while the request is pending.
    OpenShareRequestBody:
      type: object
      required: [transactionType, secretId, label, recipientKey, secretCreatedAt, senderSignature]
      properties:
        transactionType:
          $ref: '#/components/schemas/ShareTransactionType'
        secretId:
          $ref: '#/components/schemas/SecretId'
        label:
          type: string
          minLength: 1
        recipientKey:
          $ref: '#/components/schemas/PublicKey'
        secretCreatedAt:
          type: string
          format: date-time
        shareId:
          type: string
          format: uuid
          description: >
            Required for `retrieval` and `removal` requests: the `id` of the originating `deposit` row.
            Must be omitted or null for `deposit` and `inventory` requests.
        ciphertext:
          type: string
          format: byte
          description: >
            Required for `deposit` requests: the encrypted share bytes.
            Must be omitted for `retrieval`, `removal`, and `inventory` requests.
        k:
          type: integer
          description: >
            Required for `deposit` and `inventory` requests (`2 <= k <= n <= 255`,
            `400` otherwise). Must be omitted for `retrieval`/`removal`.
        n:
          type: integer
          description: >
            Required for `deposit` and `inventory` requests (`2 <= k <= n <= 255`,
            `400` otherwise). Must be omitted for `retrieval`/`removal`.
        mimeType:
          type: string
          minLength: 1
          description: >
            Required for `deposit` and `inventory` requests; must be non-blank and free of control
            characters (`400` otherwise). Must be omitted for `retrieval`/`removal`. Not validated
            against the payload — see `ShareRequest.mimeType`.
        senderSignature:
          type: string
          description: >
            Base64url Ed25519 signature by the caller over this request's canonical bytes —
            see "Payload signatures" above. Returns `400` if it doesn't verify.
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
paths:
  /share-requests:
    post:
      operationId: openShareRequest
      summary: Open a share request (deposit, retrieval, removal, or inventory)
      description: >
        Opens any share request type. The authenticated caller becomes `senderKey`.

        **deposit**: Sender deposits an encrypted share for the recipient. `ciphertext` and
        `k`/`n` are required. At most one non-denied deposit may exist per (secretId, recipientKey) pair.

        **retrieval / removal**: Sender opens a consent request. `shareId` (the originating deposit id)
        should be supplied so the recipient can look up the share in their local storage.
        At most one pending request of each type may exist per (secretId, senderKey, recipientKey) triple.

        **inventory**: A holder pushes a metadata-only report about a share back to its
        owner. `k`/`n` are required; `ciphertext` and `shareId` must be omitted. Not consent-gated
        — the row is created directly in state `approved`, and there is no conflict check (a
        holder may push repeatedly; the relay stays a dumb, idempotent mailbox).
      tags: [ShareRequests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpenShareRequestBody'
      responses:
        '201':
          description: Request opened.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareRequest'
        '400':
          description: Missing required field (e.g. ciphertext absent for deposit), or senderSignature missing/invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: An active deposit or a pending request of this type already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: >
            The request body exceeds 1 MiB, or its ciphertext exceeds 512 KiB. These are the
            relay's own storage bounds, deliberately looser than any client's own limit on the
            size of a secret.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    get:
      operationId: listShareRequests
      summary: List share requests visible to the caller
      description: |
        - `role=sender`    — requests the caller opened (to poll for resolution)
        - `role=recipient` — requests directed at the caller (pending approvals)

        Optionally filter by `type` and/or `state`.
      tags: [ShareRequests]
      parameters:
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [sender, recipient]
        - name: type
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ShareTransactionType'
          description: Filter to requests of this type.
        - name: state
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ShareRequestState'
          description: Filter to requests in this state.
      responses:
        '200':
          description: Share request list.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ShareRequest'

    delete:
      operationId: deleteShareRequests
      summary: Recipient-initiated bulk delete
      description: >
        Deletes all share request rows where the authenticated caller is the recipient.
        Optionally scoped to a specific sender and/or secret.
      tags: [ShareRequests]
      parameters:
        - name: senderKey
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PublicKey'
          description: Limit deletion to rows from this sender.
        - name: secretId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SecretId'
          description: Limit deletion to rows for this secret.
      responses:
        '204':
          description: Rows deleted (or nothing matched — idempotent).
  /share-requests/withdraw:
    post:
      operationId: withdrawShareRequests
      summary: Recipient-initiated unilateral withdrawal
      description: >
        Flips matching `approved` `deposit` rows where the authenticated caller is the recipient
        to `withdrawn`, instead of deleting them — a best-effort, fire-and-forget tombstone so the
        sender's next poll can observe that the holder stopped holding the share. Optionally
        scoped to a specific sender and/or secret. Rows in any other state (e.g. still `pending`,
        or already `denied`) are left untouched. No request body.
      tags: [ShareRequests]
      parameters:
        - name: senderKey
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/PublicKey'
          description: Limit withdrawal to deposit rows from this sender.
        - name: secretId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SecretId'
          description: Limit withdrawal to the deposit row for this secret.
      responses:
        '204':
          description: Matching rows withdrawn (or nothing matched — idempotent).
  /share-requests/{requestId}:
    get:
      operationId: getShareRequest
      summary: Get a single share request
      description: >
        Accessible by both the sender and the recipient of the request.
        When a `deposit` has been approved and the ciphertext delivered, `ciphertext` is null.
        When a `retrieval` has been approved, `ciphertext` contains the share bytes until deleted.
      tags: [ShareRequests]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Share request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareRequest'
        '403':
          description: Caller is neither the sender nor the recipient of this request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    patch:
      operationId: respondToShareRequest
      summary: Approve or deny a pending share request (recipient only)
      description: >
        Called by the recipient to approve or deny a pending request.

        - Approving a `deposit`: the relay delivers `ciphertext` to the recipient in the response body
          and clears it from the relay (one-time delivery). No `ciphertext` needed in the request body.
        - Approving a `retrieval`: the recipient must supply `ciphertext` (the share bytes from their
          local storage) in the request body. Returns `400` if absent.
        - Approving a `removal`: the relay cascades a delete of all rows for the same
          (secretId, senderKey, recipientKey) triple. No `ciphertext` involved.
        - Any denial: records `state=denied` and `respondedAt`. No ciphertext exchanged.
      tags: [ShareRequests]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [state, recipientSignature]
              properties:
                state:
                  type: string
                  enum: [approved, denied]
                ciphertext:
                  type: string
                  format: byte
                  description: Required when approving a `retrieval` request.
                recipientSignature:
                  type: string
                  description: >
                    Base64url Ed25519 signature by the caller over this response's canonical
                    bytes — see "Payload signatures" above. Required unconditionally, including
                    denials. Returns `400` if it doesn't verify.
      responses:
        '200':
          description: Decision recorded. For `deposit` approval, includes `ciphertext`.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShareRequest'
        '400':
          description: Ciphertext required (retrieval approval) but absent, or recipientSignature missing/invalid.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Caller is not the recipient of this request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Request is no longer pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: >
            The request body exceeds 1 MiB, or the ciphertext supplied with a retrieval approval
            exceeds 512 KiB.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      operationId: deleteShareRequest
      summary: Delete a single share request
      description: >
        Accessible by both the sender and the recipient.
        When deleting a `deposit` row, the relay also cascades deletion of all associated
        `retrieval` and `removal` rows for the same (secretId, senderKey, recipientKey) triple.
      tags: [ShareRequests]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Request row deleted.
        '403':
          description: Caller is neither the sender nor the recipient.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Request not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...